requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
numpy==2.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
typing-extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
